refactor(useFetchDocumentPath): Colocate with hooks
