+++ packages/cozy-sharing/src/components/ShareModal/EditableSharingModal.jsx
@@ -4,10 +4,10 @@ import React from 'react'
 import { useClient, useQueryAll } from 'cozy-client'
 
 import SharingContext from '../../context'
+import { useFetchDocumentPath } from '../../hooks/useFetchDocumentPath'
 import { Contact } from '../../models'
 import { buildContactsQuery, buildGroupsQuery } from '../../queries/queries'
 import { default as DumbShareModal } from '../ShareModal'
-import { useFetchDocumentPath } from '../useFetchDocumentPath'
 
 export const EditableSharingModal = ({ document, ...rest }) => {
   const client = useClient()
