+++ packages/cozy-sharing/src/index.jsx
@@ -7,6 +7,8 @@ export default SharingProvider
 export { SharingContext, withLocales }
 
 export { useSharingContext } from './hooks/useSharingContext'
+export { useFetchDocumentPath } from './hooks/useFetchDocumentPath'
+
 export { SharedDocument } from './SharedDocument'
 export { SharedStatus } from './SharedStatus'
 export { SharedBadge } from './SharedBadge'
@@ -16,7 +18,6 @@ export { SharedRecipientsList } from './SharedRecipientsList'
 export { ShareButton } from './ShareButton'
 export { ShareModal } from './components/ShareModal/ShareModal'
 export { RefreshableSharings } from './RefreshableSharings'
-export { useFetchDocumentPath } from './components/useFetchDocumentPath'
 export { CozyPassFingerprintDialogContent } from './components/CozyPassFingerprintDialogContent'
 export { SharingBannerPlugin } from './SharingBanner'
 export { useSharingInfos } from './SharingBanner/hooks/useSharingInfos'
